+++ src/components/book/BookGrid.tsx
@@ -7,7 +7,7 @@ export default function BookGrid({ books }: Props) {
   return (
     <ul className="grid gap-2 grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
       {books?.map(book => (
-        <li key={book.id} className="mb-10">
+        <li key={book.id} className="mb-20">
           <BookCard book={book} />
         </li>
       ))}
